+++ rust/operator-binary/src/crd/listener.rs
@@ -4,7 +4,9 @@ use std::{
 };
 
 use snafu::{OptionExt, Snafu};
-use stackable_operator::{kube::ResourceExt, utils::cluster_info::KubernetesClusterInfo};
+use stackable_operator::{
+    kube::ResourceExt, role_utils::RoleGroupRef, utils::cluster_info::KubernetesClusterInfo,
+};
 use strum::{EnumDiscriminants, EnumString};
 
 use crate::crd::{STACKABLE_LISTENER_BROKER_DIR, security::KafkaTlsSecurity, v1alpha1};
@@ -170,10 +172,14 @@ impl Display for KafkaListener {
 pub fn get_kafka_listener_config(
     kafka: &v1alpha1::KafkaCluster,
     kafka_security: &KafkaTlsSecurity,
-    object_name: &str,
+    rolegroup_ref: &RoleGroupRef<v1alpha1::KafkaCluster>,
     cluster_info: &KubernetesClusterInfo,
 ) -> Result<KafkaListenerConfig, KafkaListenerError> {
-    let pod_fqdn = pod_fqdn(kafka, object_name, cluster_info)?;
+    let pod_fqdn = pod_fqdn(
+        kafka,
+        &rolegroup_ref.rolegroup_headless_service_name(),
+        cluster_info,
+    )?;
     let mut listeners = vec![];
     let mut advertised_listeners = vec![];
     let mut listener_security_protocol_map: BTreeMap<KafkaListenerName, KafkaListenerProtocol> =
@@ -334,12 +340,11 @@ pub fn node_port_cmd(directory: &str, port_name: &str) -> String {
 
 pub fn pod_fqdn(
     kafka: &v1alpha1::KafkaCluster,
-    object_name: &str,
+    sts_service_name: &str,
     cluster_info: &KubernetesClusterInfo,
 ) -> Result<String, KafkaListenerError> {
     Ok(format!(
-        "$POD_NAME.{object_name}.{namespace}.svc.{cluster_domain}",
-        object_name = object_name,
+        "$POD_NAME.{sts_service_name}.{namespace}.svc.{cluster_domain}",
         namespace = kafka.namespace().context(ObjectHasNoNamespaceSnafu)?,
         cluster_domain = cluster_info.cluster_domain
     ))
@@ -354,7 +359,7 @@ mod tests {
     };
 
     use super::*;
-    use crate::crd::authentication::ResolvedAuthenticationClasses;
+    use crate::crd::{authentication::ResolvedAuthenticationClasses, role::KafkaRole};
 
     fn default_cluster_info() -> KubernetesClusterInfo {
         KubernetesClusterInfo {
@@ -364,9 +369,6 @@ mod tests {
 
     #[test]
     fn test_get_kafka_listeners_config() {
-        let object_name = "simple-kafka-broker-default";
-        let cluster_info = default_cluster_info();
-
         let kafka_cluster = r#"
         apiVersion: kafka.stackable.tech/v1alpha1
         kind: KafkaCluster
@@ -400,9 +402,12 @@ mod tests {
             "internalTls".to_string(),
             Some("tls".to_string()),
         );
-
+        let cluster_info = default_cluster_info();
+        // "simple-kafka-broker-default"
+        let rolegroup_ref = kafka.rolegroup_ref(&KafkaRole::Broker, "default");
         let config =
-            get_kafka_listener_config(&kafka, &kafka_security, object_name, &cluster_info).unwrap();
+            get_kafka_listener_config(&kafka, &kafka_security, &rolegroup_ref, &cluster_info)
+                .unwrap();
 
         assert_eq!(
             config.listeners(),
@@ -428,7 +433,12 @@ mod tests {
                     kafka_security.client_port_name()
                 ),
                 internal_name = KafkaListenerName::Internal,
-                internal_host = pod_fqdn(&kafka, object_name, &cluster_info).unwrap(),
+                internal_host = pod_fqdn(
+                    &kafka,
+                    &rolegroup_ref.rolegroup_headless_service_name(),
+                    &cluster_info
+                )
+                .unwrap(),
                 internal_port = kafka_security.internal_port(),
             )
         );
@@ -454,7 +464,8 @@ mod tests {
             Some("tls".to_string()),
         );
         let config =
-            get_kafka_listener_config(&kafka, &kafka_security, object_name, &cluster_info).unwrap();
+            get_kafka_listener_config(&kafka, &kafka_security, &rolegroup_ref, &cluster_info)
+                .unwrap();
 
         assert_eq!(
             config.listeners(),
@@ -480,7 +491,12 @@ mod tests {
                     kafka_security.client_port_name()
                 ),
                 internal_name = KafkaListenerName::Internal,
-                internal_host = pod_fqdn(&kafka, object_name, &cluster_info).unwrap(),
+                internal_host = pod_fqdn(
+                    &kafka,
+                    &rolegroup_ref.rolegroup_headless_service_name(),
+                    &cluster_info
+                )
+                .unwrap(),
                 internal_port = kafka_security.internal_port(),
             )
         );
@@ -505,7 +521,8 @@ mod tests {
         );
 
         let config =
-            get_kafka_listener_config(&kafka, &kafka_security, object_name, &cluster_info).unwrap();
+            get_kafka_listener_config(&kafka, &kafka_security, &rolegroup_ref, &cluster_info)
+                .unwrap();
 
         assert_eq!(
             config.listeners(),
@@ -531,7 +548,12 @@ mod tests {
                     kafka_security.client_port_name()
                 ),
                 internal_name = KafkaListenerName::Internal,
-                internal_host = pod_fqdn(&kafka, object_name, &cluster_info).unwrap(),
+                internal_host = pod_fqdn(
+                    &kafka,
+                    &rolegroup_ref.rolegroup_headless_service_name(),
+                    &cluster_info
+                )
+                .unwrap(),
                 internal_port = kafka_security.internal_port(),
             )
         );
@@ -552,9 +574,6 @@ mod tests {
 
     #[test]
     fn test_get_kafka_kerberos_listeners_config() {
-        let object_name = "simple-kafka-broker-default";
-        let cluster_info = default_cluster_info();
-
         let kafka_cluster = r#"
         apiVersion: kafka.stackable.tech/v1alpha1
         kind: KafkaCluster
@@ -587,9 +606,12 @@ mod tests {
             "tls".to_string(),
             Some("tls".to_string()),
         );
-
+        let cluster_info = default_cluster_info();
+        // "simple-kafka-broker-default"
+        let rolegroup_ref = kafka.rolegroup_ref(&KafkaRole::Broker, "default");
         let config =
-            get_kafka_listener_config(&kafka, &kafka_security, object_name, &cluster_info).unwrap();
+            get_kafka_listener_config(&kafka, &kafka_security, &rolegroup_ref, &cluster_info)
+                .unwrap();
 
         assert_eq!(
             config.listeners(),
@@ -618,7 +640,12 @@ mod tests {
                     kafka_security.client_port_name()
                 ),
                 internal_name = KafkaListenerName::Internal,
-                internal_host = pod_fqdn(&kafka, object_name, &cluster_info).unwrap(),
+                internal_host = pod_fqdn(
+                    &kafka,
+                    &rolegroup_ref.rolegroup_headless_service_name(),
+                    &cluster_info
+                )
+                .unwrap(),
                 internal_port = kafka_security.internal_port(),
                 bootstrap_name = KafkaListenerName::Bootstrap,
                 bootstrap_host = node_address_cmd(STACKABLE_LISTENER_BROKER_DIR),