+++ rust/operator-binary/src/resource/statefulset.rs
@@ -23,8 +23,8 @@ use stackable_operator::{
             apps::v1::{StatefulSet, StatefulSetSpec, StatefulSetUpdateStrategy},
             core::v1::{
                 ConfigMapKeySelector, ConfigMapVolumeSource, ContainerPort, EnvVar, EnvVarSource,
-                ExecAction, Lifecycle, LifecycleHandler, ObjectFieldSelector, PodSpec, Probe,
-                ServiceAccount, SleepAction, TCPSocketAction, Volume,
+                ExecAction, ObjectFieldSelector, PodSpec, Probe, ServiceAccount, TCPSocketAction,
+                Volume,
             },
         },
         apimachinery::pkg::{apis::meta::v1::LabelSelector, util::intstr::IntOrString},
@@ -161,7 +161,7 @@ pub enum Error {
 /// The broker rolegroup [`StatefulSet`] runs the rolegroup, as configured by the administrator.
 ///
 /// The [`Pod`](`stackable_operator::k8s_openapi::api::core::v1::Pod`)s are accessible through the corresponding
-/// [`Service`](`stackable_operator::k8s_openapi::api::core::v1::Service`) from [`build_rolegroup_service`](`crate::resource::service::build_rolegroup_service`).
+/// [`Service`](`stackable_operator::k8s_openapi::api::core::v1::Service`) from [`build_rolegroup_service`](`crate::resource::service::build_rolegroup_headless_service`).
 #[allow(clippy::too_many_arguments)]
 pub fn build_broker_rolegroup_statefulset(
     kafka: &v1alpha1::KafkaCluster,
@@ -285,13 +285,9 @@ pub fn build_broker_rolegroup_statefulset(
         ..EnvVar::default()
     });
 
-    let kafka_listeners = get_kafka_listener_config(
-        kafka,
-        kafka_security,
-        &rolegroup_ref.object_name(),
-        cluster_info,
-    )
-    .context(InvalidKafkaListenersSnafu)?;
+    let kafka_listeners =
+        get_kafka_listener_config(kafka, kafka_security, rolegroup_ref, cluster_info)
+            .context(InvalidKafkaListenersSnafu)?;
 
     let cluster_id = kafka.cluster_id().context(ClusterIdMissingSnafu)?;
 
@@ -547,7 +543,7 @@ pub fn build_broker_rolegroup_statefulset(
                 ),
                 ..LabelSelector::default()
             },
-            service_name: Some(rolegroup_ref.object_name()),
+            service_name: Some(rolegroup_ref.rolegroup_headless_service_name()),
             template: pod_template,
             volume_claim_templates: Some(pvcs),
             ..StatefulSetSpec::default()
@@ -621,8 +617,8 @@ pub fn build_controller_rolegroup_statefulset(
     });
 
     env.push(EnvVar {
-        name: "ROLEGROUP_REF".to_string(),
-        value: Some(rolegroup_ref.object_name()),
+        name: "ROLEGROUP_HEADLESS_SERVICE_NAME".to_string(),
+        value: Some(rolegroup_ref.rolegroup_headless_service_name()),
         ..EnvVar::default()
     });
 
@@ -632,13 +628,9 @@ pub fn build_controller_rolegroup_statefulset(
         ..EnvVar::default()
     });
 
-    let kafka_listeners = get_kafka_listener_config(
-        kafka,
-        kafka_security,
-        &rolegroup_ref.object_name(),
-        cluster_info,
-    )
-    .context(InvalidKafkaListenersSnafu)?;
+    let kafka_listeners =
+        get_kafka_listener_config(kafka, kafka_security, rolegroup_ref, cluster_info)
+            .context(InvalidKafkaListenersSnafu)?;
 
     cb_kafka
         .image_from_product_image(resolved_product_image)
@@ -658,6 +650,7 @@ pub fn build_controller_rolegroup_statefulset(
             kafka_security,
             &resolved_product_image.product_version,
         )])
+        .add_env_var("PRE_STOP_CONTROLLER_SLEEP_SECONDS", "10")
         .add_env_var(
             "EXTRA_ARGS",
             kafka_role
@@ -756,19 +749,7 @@ pub fn build_controller_rolegroup_statefulset(
         )
         .context(AddVolumesAndVolumeMountsSnafu)?;
 
-    // Currently, Controllers shutdown very fast, too fast in most times (flakyness) for the Brokers
-    // to off load properly. The Brokers then try to connect to any controllers until the
-    // `gracefulShutdownTimeout` is reached and the pod is finally killed.
-    // The `pre-stop` hook will delay the kill signal to the Controllers to provide the Brokers more
-    // time to offload data.
-    let mut kafka_container = cb_kafka.build();
-    kafka_container.lifecycle = Some(Lifecycle {
-        pre_stop: Some(LifecycleHandler {
-            sleep: Some(SleepAction { seconds: 10 }),
-            ..Default::default()
-        }),
-        ..Default::default()
-    });
+    let kafka_container = cb_kafka.build();
 
     pod_builder
         .metadata(metadata)
@@ -875,7 +856,7 @@ pub fn build_controller_rolegroup_statefulset(
                 ),
                 ..LabelSelector::default()
             },
-            service_name: Some(rolegroup_ref.object_name()),
+            service_name: Some(rolegroup_ref.rolegroup_headless_service_name()),
             template: pod_template,
             volume_claim_templates: Some(merged_config.resources().storage.build_pvcs()),
             ..StatefulSetSpec::default()